Separate into domains and models
